# KuryeNet/docs/openapi.yaml
openapi: 3.0.3
info:
  title: KuryeNet API
  version: 0.0.1
servers:
  - url: http://localhost:8585
paths:
...
  /api/orders/getByOrderAddressId:
    get:
      summary: getByOrderAddress_AddressId
      parameters:
        - in: query
          name: addressId
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /api/orders/getByOrderId:
    get:
      summary: getByOrderId
      parameters:
        - in: query
          name: orderId
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/orders/getByOrderNumber:
    get:
      summary: getByOrderNumber
      parameters:
        - in: query
          name: orderNumber
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/orders/getByOrderStatus:
    get:
      summary: getByOrderStatus
      parameters:
        - in: query
          name: orderStatus
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/orders/getByOrderType:
    get:
      summary: getByOrderType
      parameters:
        - in: query
          name: orderType
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /api/orders/getallByPage:
    get:
      summary: getAll
      parameters:
        - in: query
          name: pageNo
          schema: { type: string }
        - in: query
          name: pageSize
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /api/providers/getByAddressId:
    get:
      summary: getByAddress_AddressId
      parameters:
        - in: query
          name: addressId
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/providers/getByProviderCity:
    get:
      summary: getByProviderAddress_City
      parameters:
        - in: query
          name: city
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/providers/getByProviderCityAndDistrict:
    get:
      summary: getByProviderAddress_CityAndProviderAddress_District
      parameters:
        - in: query
          name: city
          schema: { type: string }
        - in: query
          name: district
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/providers/getByProviderDistrict:
    get:
      summary: getByProviderAddress_District
      parameters:
        - in: query
          name: district
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/providers/getByProviderId:
    get:
      summary: getByProviderId
      parameters:
        - in: query
          name: providerId
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/providers/getByProviderMersisNo:
    get:
      summary: getByProviderMersisNo
      parameters:
        - in: query
          name: providerMersisNo
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/providers/getByProviderName:
    get:
      summary: getByProviderName
      parameters:
        - in: query
          name: providerName
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/providers/getByProviderType:
    get:
      summary: getByProviderType
      parameters:
        - in: query
          name: providerType
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /api/providers/getallByPage:
    get:
      summary: getAll
      parameters:
        - in: query
          name: pageNo
          schema: { type: string }
        - in: query
          name: pageSize
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /api/roles/getByRoleId:
    get:
      summary: getByRoleId
      parameters:
        - in: query
          name: roleId
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/roles/getByRoleName:
    get:
      summary: getByRoleName
      parameters:
        - in: query
          name: roleName
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /api/usersroles/getByRoleName:
    get:
      summary: getByRoleName
      parameters:
        - in: query
          name: roleName
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/usersroles/getByUserEmail:
    get:
      summary: getByUserEmail
      parameters:
        - in: query
          name: userEmail
          schema: { type: string }
      responses:
        '200': { description: OK }
...